Fix UB in fixed_size_queue and move tests to the module itself
