+++ core/src/utils/fixed_size_queue.rs
@@ -1,12 +1,10 @@
 use core::{ops::{IndexMut, Index}, ptr};
 
-use super::global_static_alloctor::static_alloc;
+use super::global_static_alloctor::static_alloc_ptr;
 
-pub struct FixedSizeQueue<T:Copy + 'static, const SIZE:usize>{
-    // this uses the internal static allocator hope its works
-    _data: &'static mut [T;SIZE],   // This field is not use directly only through pointers aquired in the new() function
-    end_alloc_pointer: *const T,
-    start_alloc_pointer: *const T,
+pub struct FixedSizeQueue<T, const SIZE:usize>{
+    end_alloc_pointer: *mut T,
+    start_alloc_pointer: *mut T,
     data_pointer: *mut T,
     base_data_pointer: *mut T,
     length: usize,
@@ -14,9 +12,8 @@ pub struct FixedSizeQueue<T:Copy + 'static, const SIZE:usize>{
 
 impl<T:Copy + Default, const SIZE:usize> FixedSizeQueue<T, SIZE>{
     pub fn new()->Self{
-        let data = static_alloc([T::default(); SIZE]);
+        let data_ptr = static_alloc_ptr(SIZE).as_ptr();
         let mut s = Self{
-            _data: data,
             length:0,
             base_data_pointer: ptr::null_mut(),
             data_pointer: ptr::null_mut(),
@@ -24,10 +21,12 @@ impl<T:Copy + Default, const SIZE:usize> FixedSizeQueue<T, SIZE>{
             start_alloc_pointer: ptr::null_mut(),
         };
 
-        s.base_data_pointer = s._data.as_mut_ptr();
-        s.data_pointer = s._data.as_mut_ptr();
-        s.start_alloc_pointer = s._data.as_ptr();
-        unsafe{s.end_alloc_pointer = s._data.as_ptr().add(SIZE)};
+        s.base_data_pointer = data_ptr;
+        s.data_pointer = data_ptr;
+        s.start_alloc_pointer = data_ptr;
+
+        // SAFETY: Adding the len of the buffer to its base address resulting in off by 1 ptr which is safe
+        unsafe{s.end_alloc_pointer = data_ptr.add(SIZE)};
 
         return s;
     }
@@ -35,8 +34,8 @@ impl<T:Copy + Default, const SIZE:usize> FixedSizeQueue<T, SIZE>{
     pub fn push(&mut self, t:T){
         if self.length < SIZE{
             unsafe{
-                if self.data_pointer == self.end_alloc_pointer as *mut T{
-                    self.data_pointer = self.start_alloc_pointer as *mut T;
+                if self.data_pointer == self.end_alloc_pointer {
+                    self.data_pointer = self.start_alloc_pointer;
                 }
                 *self.data_pointer = t;
                 self.data_pointer = self.data_pointer.add(1);
@@ -53,8 +52,8 @@ impl<T:Copy + Default, const SIZE:usize> FixedSizeQueue<T, SIZE>{
             unsafe{
                 let t = *self.base_data_pointer;
                 self.base_data_pointer = self.base_data_pointer.add(1);
-                if self.base_data_pointer == self.end_alloc_pointer as *mut T{
-                    self.base_data_pointer = self.start_alloc_pointer as *mut T;
+                if self.base_data_pointer == self.end_alloc_pointer {
+                    self.base_data_pointer = self.start_alloc_pointer;
                 }
 
                 self.length -= 1;
@@ -67,8 +66,8 @@ impl<T:Copy + Default, const SIZE:usize> FixedSizeQueue<T, SIZE>{
 
     pub fn clear(&mut self){
         self.length = 0;
-        self.data_pointer = self.start_alloc_pointer as *mut T;
-        self.base_data_pointer = self.start_alloc_pointer as *mut T;
+        self.data_pointer = self.start_alloc_pointer;
+        self.base_data_pointer = self.start_alloc_pointer;
     }
 
     pub fn len(&self)->usize{
@@ -77,20 +76,20 @@ impl<T:Copy + Default, const SIZE:usize> FixedSizeQueue<T, SIZE>{
 
     pub fn fill(&mut self, value:&[T;SIZE]){
         unsafe{
-            self.base_data_pointer = self.start_alloc_pointer as *mut T;
+            self.base_data_pointer = self.start_alloc_pointer;
             ptr::copy_nonoverlapping(value.as_ptr(), self.base_data_pointer, SIZE);
             self.length = SIZE;
-            self.data_pointer = self.end_alloc_pointer as *mut T;
+            self.data_pointer = self.end_alloc_pointer;
         }
     }
 }
 
 impl<T:Copy, const SIZE:usize> FixedSizeQueue<T, SIZE>{
     #[inline] 
-    fn get_index_ptr(&self, index:usize)->*const T{
+    fn get_index_ptr(&self, index:usize)->*mut T{
         if index < self.length{
             unsafe{
-                if self.base_data_pointer.add(index) >= self.end_alloc_pointer as *mut T{
+                if self.base_data_pointer.add(index) >= self.end_alloc_pointer{
                     let wrap_offset = self.end_alloc_pointer.offset_from(self.base_data_pointer) as usize;
                     return self.start_alloc_pointer.add(index - wrap_offset);
                 }
@@ -113,6 +112,140 @@ impl<T:Copy, const SIZE:usize> Index<usize> for FixedSizeQueue<T, SIZE>{
 
 impl<T:Copy, const SIZE:usize> IndexMut<usize> for FixedSizeQueue<T, SIZE>{
     fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        unsafe{&mut *(self.get_index_ptr(index) as *mut T)}
+        unsafe{&mut *(self.get_index_ptr(index))}
+    }
+}
+
+#[cfg(test)]
+mod tests{
+    #[test]
+    fn test_fifo_ub(){
+        let mut fifo = FixedSizeQueue::<u8, 8>::new();
+        fifo.push(10);
+    }
+
+    #[test]
+    fn test_fifo(){
+        let mut fifo = FixedSizeQueue::<u8, 8>::new();
+        fifo.push(10);
+        fifo.push(22);
+
+        assert_eq!(fifo.len(), 2);
+        assert_eq!(fifo[0], 10);
+        assert_eq!(fifo[1], 22);
+
+        fifo.remove();
+        assert_eq!(fifo.len(), 1);
+        assert_eq!(fifo[0], 22);
+
+        fifo[0] = 21;
+        assert_eq!(fifo[0], 21);
+    }
+
+    #[test]
+    fn test_fifo_wrapping_around(){
+        let mut fifo = FixedSizeQueue::<u8, 3>::new();
+
+        check_push_and_remove(&mut fifo);
+        check_push_and_remove(&mut fifo);
+        check_push_and_remove(&mut fifo);
+
+        fifo.push(10);
+        fifo.push(22);
+        fifo.remove();
+        assert_eq!(fifo.len(), 1);
+        assert_eq!(fifo[0], 22);
+
+        fifo[0] = 21;
+        assert_eq!(fifo[0], 21);
+    }
+
+    fn check_push_and_remove(fifo: &mut FixedSizeQueue<u8, 3>) {
+        fifo.push(10);
+        fifo.push(22);
+        fifo.push(33);
+
+        assert_eq!(fifo.len(), 3);
+        assert_eq!(fifo[0], 10);
+        assert_eq!(fifo[1], 22);
+        assert_eq!(fifo[2], 33);
+
+        assert_eq!(fifo.remove(), 10);
+        assert_eq!(fifo.remove(), 22);
+        assert_eq!(fifo.remove(), 33);
+        assert_eq!(fifo.len(), 0);
+    }
+
+    #[test]
+    #[should_panic]
+    fn panic_on_fifo_full(){
+        let mut fifo = FixedSizeQueue::<u8, 3>::new();
+        fifo.push(1);
+        fifo.push(2);
+        fifo.push(3);
+
+        //should panic
+        fifo.push(1);
+    }
+
+    #[test]
+    #[should_panic]
+    fn panic_on_get_fifo_index_out_of_range(){
+        let mut fifo = FixedSizeQueue::<u8, 3>::new();
+        fifo.push(1);
+        fifo.push(2);
+
+        //should panic
+        let _ = fifo[2];
+    }
+
+    #[test]
+    #[should_panic]
+    fn panic_on_fifo_set_index_out_of_range(){
+        let mut fifo = FixedSizeQueue::<u8, 3>::new();
+        fifo.push(1);
+        fifo.push(2);
+
+        //should panic
+        fifo[2] = 4;
+    }
+
+    #[test]
+    fn fill_fills_the_fifo(){
+        let mut fifo = FixedSizeQueue::<u8, 8>::new();
+        fifo.push(1);
+        fifo.push(1);
+        fifo.push(1);
+
+        fifo.remove();
+        fifo.remove();
+        
+        fifo.fill(&[0;8]);
+
+        assert_eq!(fifo.len(), 8);
+        for i in 0..8{
+            assert_eq!(fifo[i], 0);
+        }
+    }
+
+    #[test]
+    fn fifo_index_check_happyflow(){
+        let mut fifo = FixedSizeQueue::<u8, 8>::new();
+        for i in 0..8{
+            fifo.push(i);
+        }
+        for _ in 0..6{
+            fifo.remove();
+        }
+        for i in 0..6{
+            fifo.push(i);
+        }
+
+        assert_eq!(fifo[0], 6);
+        assert_eq!(fifo[1], 7);
+        assert_eq!(fifo[2], 0);
+        assert_eq!(fifo[3], 1);
+        assert_eq!(fifo[4], 2);
+        assert_eq!(fifo[5], 3);
     }
 }
